made CR controllers specific again, and utilizing generic AWS Object functions now
