+++ controllers/helpers.go
@@ -2,14 +2,24 @@ package controllers
 
 import (
 	"context"
+	"time"
 
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
+	awssdk "github.com/aws/aws-sdk-go/aws"
+	awsarn "github.com/aws/aws-sdk-go/aws/arn"
+	"github.com/aws/aws-sdk-go/aws/session"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	iamv1beta1 "github.com/redradrat/aws-iam-operator/api/v1beta1"
+	"github.com/redradrat/aws-iam-operator/aws"
 )
+
+type AWSObject interface {
+	Metadata() metav1.ObjectMeta
+	GetStatus() *iamv1beta1.AWSObjectStatus
+	RuntimeObject() runtime.Object
+}
 
 // Helper functions to check and remove string from a slice of strings.
 func containsString(slice []string, s string) bool {
@@ -31,25 +41,82 @@ func removeString(slice []string, s string) (result []string) {
 	return
 }
 
-func createRoleServiceAccount(role iamv1beta1.Role, ctx context.Context, client client.Client, ownerRef metav1.OwnerReference) error {
-	if role.Spec.CreateServiceAccount {
-		sa := v1.ServiceAccount{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      role.Name,
-				Namespace: role.Namespace,
-				Labels:    role.Labels,
-				Annotations: map[string]string{
-					"eks.amazonaws.com/role-arn": role.Status.ARN,
-				},
-				OwnerReferences: []metav1.OwnerReference{ownerRef},
-			},
-		}
+type PreFunc func(obj AWSObject, c client.Client, ctx context.Context) error
 
-		if err := client.Create(ctx, &sa); err != nil && !errors.IsAlreadyExists(err) {
-			return err
-		}
+func EmptyPreFunc(obj AWSObject, c client.Client, ctx context.Context) error {
+	return nil
+}
 
+func CreateAWSObject(obj AWSObject, ins aws.Instance, preFunc PreFunc, c client.Client, ctx context.Context, sw client.StatusWriter) error {
+
+	obj.GetStatus().State = iamv1beta1.SyncSyncState
+	obj.GetStatus().LastSyncAttempt = time.Now().Format(time.RFC822Z)
+
+	session, err := startReconciliation()
+	if err != nil {
+		return err
+	}
+
+	if err := preFunc(obj, c, ctx); err != nil {
+		return err
+	}
+	arn, err := ins.Create(session)
+	if err != nil {
+		return err
+	}
+
+	obj.GetStatus().ARN = arn.String()
+	obj.GetStatus().Message = "Successfully reconciled"
+	obj.GetStatus().State = iamv1beta1.OkSyncState
+	err = sw.Update(ctx, obj.RuntimeObject())
+	if err != nil {
+		return err
 	}
 
 	return nil
+}
+
+func DeleteAWSObject(obj AWSObject, ins aws.Instance, preFunc PreFunc, c client.Client, ctx context.Context) error {
+	obj.GetStatus().State = iamv1beta1.SyncSyncState
+	obj.GetStatus().LastSyncAttempt = time.Now().Format(time.RFC822Z)
+
+	session, err := startReconciliation()
+	if err != nil {
+		return err
+	}
+
+	arn, err := awsarn.Parse(obj.GetStatus().ARN)
+	if err != nil {
+		return err
+	}
+
+	if err := preFunc(obj, c, ctx); err != nil {
+		return err
+	}
+	if err = ins.Delete(session, arn); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func errWithStatus(obj AWSObject, err error, sw client.StatusWriter, ctx context.Context) error {
+	origerr := err
+	obj.GetStatus().Message = origerr.Error()
+	obj.GetStatus().State = iamv1beta1.ErrorSyncState
+	if err = sw.Update(ctx, obj.RuntimeObject()); err != nil {
+		return err
+	}
+	return origerr
+}
+
+func startReconciliation() (*session.Session, error) {
+	session, err := session.NewSession(&awssdk.Config{
+		Region: awssdk.String("eu-west-1")},
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return session, nil
 }
